test(picker-internal): migrate keyboard-entry
